refactor: 로그인 action 함수명 변경
